+++ src/yandex_dostavka_api/entities/item.cr
@@ -5,10 +5,10 @@ module YandexDostavkaApi
       include JSON::Serializable::Strict
 
       @[JSON::Field(key: "pickup_point")]
-      property pickup_point : Int32
+      property pickup_point : Int64
 
       @[JSON::Field(key: "droppof_point")]
-      property droppof_point : Int32
+      property droppof_point : Int64
 
       @[JSON::Field(key: "title")]
       property title : String
@@ -36,7 +36,7 @@ module YandexDostavkaApi
       property extra_id : String?
 
       # fiscalization
-      def initialize(@pickup_point : Int32 = 0, @droppof_point : Int32 = 0, @title : String = "Товары", @size : Size = Size.new(0, 0, 0), @weight : Float64 = 0.0, @quantity : Int32 = 1, cost_value : Float64 = 0.0, @extra_id : String? = nil)
+      def initialize(@pickup_point : Int64 = 0, @droppof_point : Int64 = 0, @title : String = "Товары", @size : Size = Size.new(0, 0, 0), @weight : Float64 = 0.0, @quantity : Int32 = 1, cost_value : Float64 = 0.0, @extra_id : String? = nil)
         @cost_value = cost_value.to_s
       end
